implement AAID, BB, BGS and CARMA inflections
