+++ spec/lib/carma/models/metadata_spec.rb
@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 
-RSpec.describe CARMA::Models::Submission::Metadata, type: :model do
+RSpec.describe CARMA::Models::Metadata, type: :model do
   describe '#claim_id' do
     it 'is accessible' do
       subject.claim_id = 123
@@ -14,7 +14,7 @@ RSpec.describe CARMA::Models::Submission::Metadata, type: :model do
     it 'is accessible' do
       subject.veteran = { icn: 'ABCD1234', is_veteran: true }
 
-      expect(subject.veteran).to be_instance_of(described_class::Veteran)
+      expect(subject.veteran).to be_instance_of(CARMA::Models::Veteran)
       expect(subject.veteran.icn).to eq('ABCD1234')
       expect(subject.veteran.is_veteran).to eq(true)
     end
@@ -24,7 +24,7 @@ RSpec.describe CARMA::Models::Submission::Metadata, type: :model do
     it 'is accessible' do
       subject.primary_caregiver = { icn: 'ABCD1234' }
 
-      expect(subject.primary_caregiver).to be_instance_of(described_class::Caregiver)
+      expect(subject.primary_caregiver).to be_instance_of(CARMA::Models::Caregiver)
       expect(subject.primary_caregiver.icn).to eq('ABCD1234')
     end
   end
@@ -33,7 +33,7 @@ RSpec.describe CARMA::Models::Submission::Metadata, type: :model do
     it 'is accessible' do
       subject.secondary_caregiver_one = { icn: 'ABCD1234' }
 
-      expect(subject.secondary_caregiver_one).to be_instance_of(described_class::Caregiver)
+      expect(subject.secondary_caregiver_one).to be_instance_of(CARMA::Models::Caregiver)
       expect(subject.secondary_caregiver_one.icn).to eq('ABCD1234')
     end
 
@@ -47,7 +47,7 @@ RSpec.describe CARMA::Models::Submission::Metadata, type: :model do
     it 'is accessible' do
       subject.secondary_caregiver_two = { icn: 'ABCD1234' }
 
-      expect(subject.secondary_caregiver_two).to be_instance_of(described_class::Caregiver)
+      expect(subject.secondary_caregiver_two).to be_instance_of(CARMA::Models::Caregiver)
       expect(subject.secondary_caregiver_two.icn).to eq('ABCD1234')
     end
 
@@ -60,12 +60,12 @@ RSpec.describe CARMA::Models::Submission::Metadata, type: :model do
   describe '::new' do
     it 'initializes with defaults' do
       # Should default to empty described_class::Veteran
-      expect(subject.veteran).to be_instance_of(described_class::Veteran)
+      expect(subject.veteran).to be_instance_of(CARMA::Models::Veteran)
       expect(subject.veteran.icn).to eq(nil)
       expect(subject.veteran.is_veteran).to eq(nil)
 
       # Should default to empty described_class::Caregiver
-      expect(subject.primary_caregiver).to be_instance_of(described_class::Caregiver)
+      expect(subject.primary_caregiver).to be_instance_of(CARMA::Models::Caregiver)
       expect(subject.primary_caregiver.icn).to eq(nil)
 
       # Should default to nil
