+++ spec/lib/carma/models/veteran_spec.rb
@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 
-RSpec.describe CARMA::Models::Submission::Metadata::Caregiver, type: :model do
+RSpec.describe CARMA::Models::Veteran, type: :model do
   describe '#icn' do
     it 'is accessible' do
       subject.icn = 'ABCD1234'
@@ -10,10 +10,18 @@ RSpec.describe CARMA::Models::Submission::Metadata::Caregiver, type: :model do
     end
   end
 
+  describe '#is_veteran' do
+    it 'is accessible' do
+      subject.is_veteran = true
+      expect(subject.is_veteran).to eq(true)
+    end
+  end
+
   describe '::new' do
     it 'is :is_veteran, :icn' do
-      subject = described_class.new(icn: 'ABCD1234')
+      subject = described_class.new(icn: 'ABCD1234', is_veteran: true)
       expect(subject.icn).to eq('ABCD1234')
+      expect(subject.is_veteran).to eq(true)
     end
   end
 
@@ -23,17 +31,18 @@ RSpec.describe CARMA::Models::Submission::Metadata::Caregiver, type: :model do
     end
 
     it 'sets request_payload_keys' do
-      expect(described_class.request_payload_keys).to eq([:icn])
+      expect(described_class.request_payload_keys).to eq(%i[icn is_veteran])
     end
   end
 
   describe '#to_request_payload' do
     it 'can receive :to_request_payload' do
-      instance = described_class.new icn: 'ABCD1234'
+      instance = described_class.new icn: 'ABCD1234', is_veteran: true
 
       expect(instance.to_request_payload).to eq(
         {
-          'icn' => 'ABCD1234'
+          'icn' => 'ABCD1234',
+          'isVeteran' => true
         }
       )
     end
